extract routes from model to file
